+++ test/unit/test_jinja.py
@@ -1,3 +1,4 @@
+from contextlib import contextmanager
 import pytest
 import unittest
 import yaml
@@ -5,7 +6,373 @@ import yaml
 from dbt.clients.jinja import get_rendered
 from dbt.clients.jinja import get_template
 from dbt.clients.jinja import extract_toplevel_blocks
-from dbt.exceptions import CompilationException
+from dbt.exceptions import CompilationException, JinjaRenderingException
+
+
+@contextmanager
+def returns(value):
+    yield value
+
+
+@contextmanager
+def raises(value):
+    with pytest.raises(value) as exc:
+        yield exc
+
+
+def expected_id(arg):
+    if isinstance(arg, list):
+        return '_'.join(arg)
+
+
+jinja_tests = [
+    # strings
+    (
+        '''foo: bar''',
+        returns('bar'),
+        returns('bar'),
+    ),
+    (
+        '''foo: "bar"''',
+        returns('bar'),
+        returns('bar'),
+    ),
+    (
+        '''foo: "'bar'"''',
+        returns("'bar'"),
+        returns("'bar'"),
+    ),
+    (
+        """foo: '"bar"'""",
+        returns('"bar"'),
+        returns('"bar"'),
+    ),
+    (
+        '''foo: "{{ 'bar' | as_text }}"''',
+        returns('bar'),
+        returns('bar'),
+    ),
+    (
+        '''foo: "{{ 'bar' | as_bool }}"''',
+        returns('bar'),
+        raises(JinjaRenderingException),
+    ),
+    (
+        '''foo: "{{ 'bar' | as_number }}"''',
+        returns('bar'),
+        raises(JinjaRenderingException),
+    ),
+    (
+        '''foo: "{{ 'bar' | as_native }}"''',
+        returns('bar'),
+        returns('bar'),
+    ),
+    # ints
+    (
+        '''foo: 1''',
+        returns('1'),
+        returns('1'),
+    ),
+    (
+        '''foo: "1"''',
+        returns('1'),
+        returns('1'),
+    ),
+    (
+        '''foo: "'1'"''',
+        returns("'1'"),
+        returns("'1'"),
+    ),
+    (
+        """foo: '"1"'""",
+        returns('"1"'),
+        returns('"1"'),
+    ),
+    (
+        '''foo: "{{ 1 }}"''',
+        returns('1'),
+        returns('1'),
+    ),
+    (
+        '''foo: "{{ '1' }}"''',
+        returns('1'),
+        returns('1'),
+    ),
+    (
+        '''foo: "'{{ 1 }}'"''',
+        returns("'1'"),
+        returns("'1'"),
+    ),
+    (
+        '''foo: "'{{ '1' }}'"''',
+        returns("'1'"),
+        returns("'1'"),
+    ),
+    (
+        '''foo: "{{ 1 | as_text }}"''',
+        returns('1'),
+        returns('1'),
+    ),
+    (
+        '''foo: "{{ 1 | as_bool }}"''',
+        returns('1'),
+        raises(JinjaRenderingException),
+    ),
+    (
+        '''foo: "{{ 1 | as_number }}"''',
+        returns('1'),
+        returns(1),
+    ),
+    (
+        '''foo: "{{ 1 | as_native }}"''',
+        returns('1'),
+        returns(1),
+    ),
+    (
+        '''foo: "{{ '1' | as_text }}"''',
+        returns('1'),
+        returns('1'),
+    ),
+    (
+        '''foo: "{{ '1' | as_bool }}"''',
+        returns('1'),
+        raises(JinjaRenderingException),
+    ),
+    (
+        '''foo: "{{ '1' | as_number }}"''',
+        returns('1'),
+        returns(1),
+    ),
+    (
+        '''foo: "{{ '1' | as_native }}"''',
+        returns('1'),
+        returns(1),
+    ),
+    # booleans.
+    # Note the discrepancy with true vs True: `true` is recognized by jinja but
+    # not literal_eval, but `True` is recognized by ast.literal_eval.
+    # For extra fun, yaml recognizes both.
+    # unquoted true
+    (
+        '''foo: "{{ True }}"''',
+        returns('True'),
+        returns('True'),
+    ),
+    (
+        '''foo: "{{ True | as_text }}"''',
+        returns('True'),
+        returns('True'),
+    ),
+    (
+        '''foo: "{{ True | as_bool }}"''',
+        returns('True'),
+        returns(True),
+    ),
+    (
+        '''foo: "{{ True | as_number }}"''',
+        returns('True'),
+        raises(JinjaRenderingException),
+    ),
+    (
+        '''foo: "{{ True | as_native }}"''',
+        returns('True'),
+        returns(True),
+    ),
+    # unquoted true
+    (
+        '''foo: "{{ true }}"''',
+        returns("True"),
+        returns("True"),
+    ),
+    (
+        '''foo: "{{ true | as_text }}"''',
+        returns("True"),
+        returns("True"),
+    ),
+    (
+        '''foo: "{{ true | as_bool }}"''',
+        returns("True"),
+        returns(True),
+    ),
+    (
+        '''foo: "{{ true | as_number }}"''',
+        returns("True"),
+        raises(JinjaRenderingException),
+    ),
+    (
+        '''foo: "{{ true | as_native }}"''',
+        returns("True"),
+        returns(True),
+    ),
+    (
+        '''foo: "{{ 'true' | as_text }}"''',
+        returns("true"),
+        returns("true"),
+    ),
+    # quoted 'true'
+    (
+        '''foo: "'{{ true }}'"''',
+        returns("'True'"),
+        returns("'True'"),
+    ),  # jinja true -> python True -> str(True) -> "True" -> quoted
+    (
+        '''foo: "'{{ true | as_text }}'"''',
+        returns("'True'"),
+        returns("'True'"),
+    ),
+    (
+        '''foo: "'{{ true | as_bool }}'"''',
+        returns("'True'"),
+        returns("'True'"),
+    ),
+    (
+        '''foo: "'{{ true | as_number }}'"''',
+        returns("'True'"),
+        returns("'True'"),
+    ),
+    (
+        '''foo: "'{{ true | as_native }}'"''',
+        returns("'True'"),
+        returns("'True'"),
+    ),
+    # unquoted True
+    (
+        '''foo: "{{ True }}"''',
+        returns('True'),
+        returns('True'),
+    ),
+    (
+        '''foo: "{{ True | as_text }}"''',
+        returns("True"),
+        returns("True"),
+    ),  # True -> string 'True' -> text -> str('True') -> 'True'
+    (
+        '''foo: "{{ True | as_bool }}"''',
+        returns("True"),
+        returns(True),
+    ),
+    (
+        '''foo: "{{ True | as_number }}"''',
+        returns("True"),
+        raises(JinjaRenderingException),
+    ),
+    (
+        '''foo: "{{ True | as_native }}"''',
+        returns("True"),
+        returns(True),
+    ),
+    # quoted 'True' within rendering
+    (
+        '''foo: "{{ 'True' | as_text }}"''',
+        returns("True"),
+        returns("True"),
+    ),
+    # 'True' -> string 'True' -> text -> str('True') -> 'True'
+    (
+        '''foo: "{{ 'True' | as_bool }}"''',
+        returns('True'),
+        returns(True),
+    ),
+    # quoted 'True' outside rendering
+    (
+        '''foo: "'{{ True }}'"''',
+        returns("'True'"),
+        returns("'True'"),
+    ),
+    (
+        '''foo: "'{{ True | as_bool }}'"''',
+        returns("'True'"),
+        returns("'True'"),
+    ),
+    # yaml turns 'yes' into a boolean true
+    (
+        '''foo: yes''',
+        returns('True'),
+        returns('True'),
+    ),
+    (
+        '''foo: "yes"''',
+        returns('yes'),
+        returns('yes'),
+    ),
+    # concatenation
+    (
+        '''foo: "{{ (a_int + 100) | as_native }}"''',
+        returns('200'),
+        returns(200),
+    ),
+    (
+        '''foo: "{{ (a_str ~ 100) | as_native }}"''',
+        returns('100100'),
+        returns(100100),
+    ),
+    (
+        '''foo: "{{( a_int ~ 100) | as_native }}"''',
+        returns('100100'),
+        returns(100100),
+    ),
+    # multiple nodes -> always str
+    (
+        '''foo: "{{ a_str | as_native }}{{ a_str | as_native }}"''',
+        returns('100100'),
+        returns('100100'),
+    ),
+    (
+        '''foo: "{{ a_int | as_native }}{{ a_int | as_native }}"''',
+        returns('100100'),
+        returns('100100'),
+    ),
+    (
+        '''foo: "'{{ a_int | as_native }}{{ a_int | as_native }}'"''',
+        returns("'100100'"),
+        returns("'100100'"),
+    ),
+    (
+        '''foo:''',
+        returns('None'),
+        returns('None'),
+    ),
+    (
+        '''foo: null''',
+        returns('None'),
+        returns('None'),
+    ),
+    (
+        '''foo: ""''',
+        returns(''),
+        returns(''),
+    ),
+    (
+        '''foo: "{{ '' | as_native }}"''',
+        returns(''),
+        returns(''),
+    ),
+    # very annoying, but jinja 'none' is yaml 'null'.
+    (
+        '''foo: "{{ none | as_native }}"''',
+        returns('None'),
+        returns(None),
+    ),
+]
+
+
+@pytest.mark.parametrize(
+    'value,text_expectation,native_expectation',
+    jinja_tests,
+    ids=expected_id
+)
+def test_jinja_rendering(value, text_expectation, native_expectation):
+    foo_value = yaml.safe_load(value)['foo']
+    ctx = {
+        'a_str': '100',
+        'a_int': 100,
+        'b_str': 'hello'
+    }
+    with text_expectation as text_result:
+        assert text_result == get_rendered(foo_value, ctx, native=False)
+
+    with native_expectation as native_result:
+        assert native_result == get_rendered(foo_value, ctx, native=True)
 
 
 class TestJinja(unittest.TestCase):
@@ -17,25 +384,25 @@ class TestJinja(unittest.TestCase):
         self.assertEqual(mod.my_dict, {'a': 1})
 
     def test_regular_render(self):
-        s = '{{ "some_value" }}'
+        s = '{{ "some_value" | as_native }}'
         value = get_rendered(s, {}, native=False)
         assert value == 'some_value'
-        s = '{{ 1991 }}'
+        s = '{{ 1991 | as_native }}'
         value = get_rendered(s, {}, native=False)
         assert value == '1991'
 
         s = '{{ "some_value" | as_text }}'
-        value = get_rendered(s, {}, native=True)
+        value = get_rendered(s, {}, native=False)
         assert value == 'some_value'
         s = '{{ 1991 | as_text }}'
-        value = get_rendered(s, {}, native=True)
+        value = get_rendered(s, {}, native=False)
         assert value == '1991'
 
     def test_native_render(self):
-        s = '{{ "some_value" }}'
+        s = '{{ "some_value" | as_native }}'
         value = get_rendered(s, {}, native=True)
         assert value == 'some_value'
-        s = '{{ 1991 }}'
+        s = '{{ 1991 | as_native }}'
         value = get_rendered(s, {}, native=True)
         assert value == 1991
 
@@ -413,65 +780,3 @@ if_you_do_this_you_are_awful = '''
 hi
 {% endmaterialization %}
 '''
-
-
-native_expected_behaviors = [
-    # strings
-    ('''foo: bar''', 'bar'),
-    ('''foo: "bar"''', 'bar'),
-    ('''foo: "'bar'"''', "'bar'"),
-    ("""foo: '"bar"'""", '"bar"'),
-    # ints
-    ('''foo: 1''', 1),
-    ('''foo: "1"''', 1),
-    ('''foo: "'1'"''', "'1'"),
-    ('''foo: "{{ 1 }}"''', 1),
-    ('''foo: "{{ '1' }}"''', 1),
-    ('''foo: "'{{ 1 }}'"''', "'1'"),
-    ('''foo: "'{{ '1' }}'"''', "'1'"),
-    ('''foo: "{{ 1 | as_text }}"''', '1'),
-    ('''foo: "{{ '1' | as_text }}"''', '1'),
-    # booleans.
-    # Note the discrepancy with true vs True: `true` is recognized by jinja but
-    # not literal_eval, but `True` is recognized by ast.literal_eval.
-    # For extra fun, yaml recognizes both.
-    ('''foo: "{{ true }}"''', True),
-    ('''foo: "{{ 'true' }}"''', 'true'),
-    ('''foo: "'{{ true }}'"''', "'True'"),
-    ('''foo: "{{ true | as_text }}"''', "True"),  # true -> boolean True -> text -> str(True) -> 'True'
-    ('''foo: "{{ 'true' | as_text }}"''', "true"),  # 'true' -> string 'true' -> text -> str('true') -> 'true'
-    ('''foo: "{{ True }}"''', True),
-    ('''foo: "{{ 'True' }}"''', True),
-    ('''foo: "'{{ True }}'"''', "'True'"),
-    ('''foo: "{{ True | as_text }}"''', "True"),  # True -> string 'True' -> text -> str('True') -> 'True'
-    ('''foo: "{{ 'True' | as_text }}"''', "True"),  # 'True' -> string 'True' -> text -> str('True') -> 'True'
-    ('''foo: yes''', True),  # yaml turns 'yes' into a boolean true
-    ('''foo: "yes"''', "yes"),
-    # concatenation
-    ('''foo: "{{ a_int + 100 }}"''', 200),
-    ('''foo: "{{ a_str ~ 100 }}"''', 100100),
-    ('''foo: "{{ a_int ~ 100 }}"''', 100100),
-    ('''foo: "{{ a_str }}{{ a_str }}"''', 100100),
-    ('''foo: "{{ a_int }}{{ a_int }}"''', 100100),
-    ('''foo: "'{{ a_int }}{{ a_int }}'"''', "'100100'"),
-
-]
-
-
-def expected_id(arg):
-    if isinstance(arg, list):
-        return '_'.join(arg)
-
-
-@pytest.mark.parametrize(
-    'inputvalue,expected', native_expected_behaviors, ids=expected_id
-)
-def test_native_rendering(inputvalue, expected):
-    # this test is pretty useless without preprocessing things in yaml.
-    value = yaml.safe_load(inputvalue)['foo']
-    ctx = {
-        'a_str': '100',
-        'a_int': 100,
-        'b_str': 'hello'
-    }
-    assert get_rendered(value, ctx, native=True) == expected